LRCI-3722 Rename package 'build' to 'bui1d' to avoid having to use 'git add -f'
