+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/dalo/BuildToBuildRunsDALO.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.build.dalo;
+package com.liferay.jethr0.bui1d.dalo;
 
-import com.liferay.jethr0.build.Build;
-import com.liferay.jethr0.build.BuildFactory;
-import com.liferay.jethr0.build.parameter.BuildParameter;
-import com.liferay.jethr0.build.parameter.BuildParameterFactory;
+import com.liferay.jethr0.bui1d.Build;
+import com.liferay.jethr0.bui1d.BuildFactory;
+import com.liferay.jethr0.bui1d.run.BuildRun;
+import com.liferay.jethr0.bui1d.run.BuildRunFactory;
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
 
@@ -19,12 +19,12 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class BuildToBuildParametersDALO
-	extends BaseEntityRelationshipDALO<Build, BuildParameter> {
+public class BuildToBuildRunsDALO
+	extends BaseEntityRelationshipDALO<Build, BuildRun> {
 
 	@Override
-	public EntityFactory<BuildParameter> getChildEntityFactory() {
-		return _buildParameterFactory;
+	public EntityFactory<BuildRun> getChildEntityFactory() {
+		return _buildRunFactory;
 	}
 
 	@Override
@@ -34,13 +34,13 @@ public class BuildToBuildParametersDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "buildToBuildParameters";
+		return "buildToBuildRuns";
 	}
 
 	@Autowired
 	private BuildFactory _buildFactory;
 
 	@Autowired
-	private BuildParameterFactory _buildParameterFactory;
+	private BuildRunFactory _buildRunFactory;
 
 }
