+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/dalo/BuildToEnvironmentsDALO.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.build.dalo;
+package com.liferay.jethr0.bui1d.dalo;
 
-import com.liferay.jethr0.build.Build;
-import com.liferay.jethr0.build.BuildFactory;
-import com.liferay.jethr0.build.run.BuildRun;
-import com.liferay.jethr0.build.run.BuildRunFactory;
+import com.liferay.jethr0.bui1d.Build;
+import com.liferay.jethr0.bui1d.BuildFactory;
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
+import com.liferay.jethr0.environment.Environment;
+import com.liferay.jethr0.environment.EnvironmentFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +19,12 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class BuildToBuildRunsDALO
-	extends BaseEntityRelationshipDALO<Build, BuildRun> {
+public class BuildToEnvironmentsDALO
+	extends BaseEntityRelationshipDALO<Build, Environment> {
 
 	@Override
-	public EntityFactory<BuildRun> getChildEntityFactory() {
-		return _buildRunFactory;
+	public EntityFactory<Environment> getChildEntityFactory() {
+		return _environmentFactory;
 	}
 
 	@Override
@@ -34,13 +34,13 @@ public class BuildToBuildRunsDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "buildToBuildRuns";
+		return "buildToEnvironments";
 	}
 
 	@Autowired
 	private BuildFactory _buildFactory;
 
 	@Autowired
-	private BuildRunFactory _buildRunFactory;
+	private EnvironmentFactory _environmentFactory;
 
 }
