+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/dalo/BuildToTasksDALO.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.build.dalo;
+package com.liferay.jethr0.bui1d.dalo;
 
-import com.liferay.jethr0.build.Build;
-import com.liferay.jethr0.build.BuildFactory;
+import com.liferay.jethr0.bui1d.Build;
+import com.liferay.jethr0.bui1d.BuildFactory;
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
-import com.liferay.jethr0.environment.Environment;
-import com.liferay.jethr0.environment.EnvironmentFactory;
+import com.liferay.jethr0.task.Task;
+import com.liferay.jethr0.task.TaskFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +19,11 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class BuildToEnvironmentsDALO
-	extends BaseEntityRelationshipDALO<Build, Environment> {
+public class BuildToTasksDALO extends BaseEntityRelationshipDALO<Build, Task> {
 
 	@Override
-	public EntityFactory<Environment> getChildEntityFactory() {
-		return _environmentFactory;
+	public EntityFactory<Task> getChildEntityFactory() {
+		return _taskFactory;
 	}
 
 	@Override
@@ -34,13 +33,13 @@ public class BuildToEnvironmentsDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "buildToEnvironments";
+		return "buildToTasks";
 	}
 
 	@Autowired
 	private BuildFactory _buildFactory;
 
 	@Autowired
-	private EnvironmentFactory _environmentFactory;
+	private TaskFactory _taskFactory;
 
 }
