+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/run/BaseBuildRun.java
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.build.run;
+package com.liferay.jethr0.bui1d.run;
 
-import com.liferay.jethr0.build.Build;
-import com.liferay.jethr0.build.parameter.BuildParameter;
+import com.liferay.jethr0.bui1d.Build;
+import com.liferay.jethr0.bui1d.parameter.BuildParameter;
 import com.liferay.jethr0.entity.BaseEntity;
 import com.liferay.jethr0.jenkins.node.JenkinsNode;
 import com.liferay.jethr0.project.Project;
